refactor: remove sending hooks
