+++ src/services/postToLens.ts
@@ -2,11 +2,18 @@ import { image, MediaImageMimeType, MediaVideoMimeType, textOnly, video } from '
 import { t } from '@lingui/macro';
 import { v4 as uuid } from 'uuid';
 
+import { queryClient } from '@/configs/queryClient.js';
 import { SocialPlatform } from '@/constants/enum.js';
 import { SITE_URL } from '@/constants/index.js';
+import { enqueueErrorMessage, enqueueSuccessMessage } from '@/helpers/enqueueMessage.js';
 import { getUserLocale } from '@/helpers/getUserLocale.js';
+import { readChars } from '@/helpers/readChars.js';
 import { LensSocialMediaProvider } from '@/providers/lens/SocialMedia.js';
 import { uploadToArweave } from '@/services/uploadToArweave.js';
+import { uploadFileToIPFS } from '@/services/uploadToIPFS.js';
+import { type CompositePost, useComposeStateStore } from '@/store/useComposeStore.js';
+import { useLensStateStore } from '@/store/useProfileStore.js';
+import type { ComposeType } from '@/types/compose.js';
 import type { MediaObject } from '@/types/index.js';
 
 interface BaseMetadata {
@@ -126,7 +133,7 @@ function createPostMetadata(baseMetadata: BaseMetadata, attachments?: Attachment
 
 export type GetPostMetaData = ReturnType<typeof createPostMetadata>;
 
-export async function publishPostForLens(
+async function publishPostForLens(
     profileId: string,
     content: string,
     images: MediaObject[],
@@ -163,7 +170,7 @@ export async function publishPostForLens(
     return post;
 }
 
-export async function commentPostForLens(
+async function commentPostForLens(
     profileId: string,
     postId: string,
     content: string,
@@ -192,7 +199,7 @@ export async function commentPostForLens(
     return LensSocialMediaProvider.commentPost(postId, `ar://${arweaveId}`, profile.signless, onMomoka);
 }
 
-export async function quotePostForLens(
+async function quotePostForLens(
     profileId: string,
     postId: string,
     content: string,
@@ -220,4 +227,108 @@ export async function quotePostForLens(
     const arweaveId = await uploadToArweave(metadata, token);
     const post = await LensSocialMediaProvider.quotePost(postId, `ar://${arweaveId}`, profile.signless, onMomoka);
     return post;
+}
+
+export async function postToLens(type: ComposeType, compositePost: CompositePost) {
+    const { chars, images, lensPostId, post, video } = compositePost;
+
+    const { currentProfile } = useLensStateStore.getState();
+    if (!currentProfile?.profileId || lensPostId) return;
+
+    const { updateImages, updateVideo, updateLensPostId } = useComposeStateStore.getState();
+
+    const uploadedImages = await Promise.all(
+        images.map(async (media) => {
+            try {
+                if (media.ipfs) return media;
+                const ipfs = await uploadFileToIPFS(media.file);
+                const patchedMedia: MediaObject = {
+                    ...media,
+                    ipfs,
+                };
+                updateImages((originImages) => {
+                    return originImages.map((x) => (x.file === media.file ? patchedMedia : x));
+                });
+                // We only care about ipfs for Lens
+                return patchedMedia;
+            } catch (err) {
+                const message = t`Failed to upload image to IPFS.`;
+                enqueueErrorMessage(message);
+                throw new Error(message);
+            }
+        }),
+    );
+    let uploadedVideo = video;
+    if (video?.file && !video.ipfs) {
+        const response = await uploadFileToIPFS(video.file);
+        if (response) {
+            uploadedVideo = {
+                ...video,
+                ipfs: response,
+            };
+            updateVideo(uploadedVideo);
+        } else {
+            const message = t`Failed to upload video to IPFS.`;
+            enqueueErrorMessage(message);
+            throw new Error(message);
+        }
+    }
+
+    if (type === 'compose') {
+        try {
+            const published = await publishPostForLens(
+                currentProfile.profileId,
+                readChars(chars),
+                uploadedImages,
+                uploadedVideo,
+            );
+            enqueueSuccessMessage(t`Posted on Lens`);
+            updateLensPostId(published.postId);
+        } catch (error) {
+            enqueueErrorMessage(t`Failed to post on Lens.`);
+            throw error;
+        }
+    } else if (type === 'reply') {
+        try {
+            if (!post) throw new Error('No post found.');
+            const comment = await commentPostForLens(
+                currentProfile.profileId,
+                post.postId,
+                readChars(chars),
+                uploadedImages,
+                uploadedVideo,
+                !!post.momoka?.proof,
+            );
+            enqueueSuccessMessage(t`Replied on Lens`);
+            updateLensPostId(comment);
+
+            queryClient.invalidateQueries({ queryKey: [post.source, 'post-detail', post.postId] });
+            queryClient.invalidateQueries({ queryKey: ['post-detail', 'comments', post.source, post.postId] });
+        } catch (error) {
+            enqueueErrorMessage(t`Failed to relay post on Lens.`);
+            throw error;
+        }
+    } else if (type === 'quote') {
+        try {
+            if (!post) throw new Error('No post found.');
+            const quote = await quotePostForLens(
+                currentProfile.profileId,
+                post.postId,
+                readChars(chars),
+                uploadedImages,
+                uploadedVideo,
+                !!post.momoka?.proof,
+            );
+            enqueueSuccessMessage(t`Posted on Lens`);
+            updateLensPostId(quote.postId);
+
+            await queryClient.setQueryData([post.source, 'post-detail', post.postId], {
+                ...post,
+                hasQuoted: true,
+            });
+        } catch (error) {
+            enqueueErrorMessage(t`Failed to quote post on Lens.`);
+            throw error;
+        }
+    }
 }
