+++ src/services/uploadToArweave.ts
@@ -2,7 +2,7 @@ import urlcat from 'urlcat';
 
 import { FIREFLY_ROOT_URL } from '@/constants/index.js';
 import { fetchJSON } from '@/helpers/fetchJSON.js';
-import type { GetPostMetaData } from '@/services/postForLens.js';
+import type { GetPostMetaData } from '@/services/postToLens.js';
 
 interface IUploadToArweaveResponse {
     data: {
